+++ Modules/Intent.Modules.Ardalis.Repositories/NugetPackages.cs
@@ -19,9 +19,9 @@ namespace Intent.Modules.Ardalis.Repositories
             NugetRegistry.Register(ArdalisSpecificationPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 9, >= 0) => new PackageVersion("9.2.0"),
-                        ( >= 8, >= 0) => new PackageVersion("9.2.0"),
-                        ( >= 2, >= 0) => new PackageVersion("9.2.0")
+                        ( >= 9, >= 0) => new PackageVersion("9.3.1"),
+                        ( >= 8, >= 0) => new PackageVersion("9.3.1"),
+                        ( >= 2, >= 0) => new PackageVersion("9.3.1")
                             .WithNugetDependency("System.Buffers", "4.6.1")
                             .WithNugetDependency("System.Memory", "4.6.3"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{ArdalisSpecificationPackageName}'"),
@@ -30,14 +30,14 @@ namespace Intent.Modules.Ardalis.Repositories
             NugetRegistry.Register(ArdalisSpecificationEntityFrameworkCorePackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 9, >= 0) => new PackageVersion("9.2.0")
-                            .WithNugetDependency("Ardalis.Specification", "9.2.0")
-                            .WithNugetDependency("Microsoft.EntityFrameworkCore", "9.0.5")
-                            .WithNugetDependency("Microsoft.EntityFrameworkCore.Relational", "9.0.5"),
-                        ( >= 8, >= 0) => new PackageVersion("9.2.0")
-                            .WithNugetDependency("Ardalis.Specification", "9.2.0")
-                            .WithNugetDependency("Microsoft.EntityFrameworkCore", "8.0.16")
-                            .WithNugetDependency("Microsoft.EntityFrameworkCore.Relational", "8.0.16"),
+                        ( >= 9, >= 0) => new PackageVersion("9.3.1")
+                            .WithNugetDependency("Ardalis.Specification", "9.3.1")
+                            .WithNugetDependency("Microsoft.EntityFrameworkCore", "9.0.8")
+                            .WithNugetDependency("Microsoft.EntityFrameworkCore.Relational", "9.0.8"),
+                        ( >= 8, >= 0) => new PackageVersion("9.3.1")
+                            .WithNugetDependency("Ardalis.Specification", "9.3.1")
+                            .WithNugetDependency("Microsoft.EntityFrameworkCore", "8.0.19")
+                            .WithNugetDependency("Microsoft.EntityFrameworkCore.Relational", "8.0.19"),
                         ( >= 7, >= 0) => new PackageVersion("8.0.0"),
                         ( >= 6, >= 0) => new PackageVersion("8.0.0"),
                         ( >= 2, >= 1) => new PackageVersion("6.1.0")