+++ Modules/Intent.Modules.Aws.Common/NugetPackages.cs
@@ -18,12 +18,12 @@ namespace Intent.Modules.Aws.Common
             NugetRegistry.Register(AWSSDKExtensionsNETCoreSetupPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 8, >= 0) => new PackageVersion("4.0.2.1")
+                        ( >= 8, >= 0) => new PackageVersion("4.0.2.2")
                             .WithNugetDependency("AWSSDK.Core", "4.0.0.18")
                             .WithNugetDependency("Microsoft.Extensions.Configuration.Abstractions", "2.0.0")
                             .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "8.0.0")
                             .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "2.0.0"),
-                        ( >= 2, >= 0) => new PackageVersion("4.0.2.1")
+                        ( >= 2, >= 0) => new PackageVersion("4.0.2.2")
                             .WithNugetDependency("AWSSDK.Core", "4.0.0.18")
                             .WithNugetDependency("Microsoft.Extensions.Configuration.Abstractions", "2.0.0")
                             .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "2.0.0")