+++ Modules/Intent.Modules.Blazor/NugetPackages.cs
@@ -20,18 +20,18 @@ namespace Intent.Modules.Blazor
             NugetRegistry.Register(MicrosoftAspNetCoreComponentsWebAssemblyPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 9, >= 0) => new PackageVersion("9.0.7")
-                            .WithNugetDependency("Microsoft.AspNetCore.Components.Web", "9.0.7")
-                            .WithNugetDependency("Microsoft.Extensions.Configuration.Binder", "9.0.7")
-                            .WithNugetDependency("Microsoft.Extensions.Configuration.Json", "9.0.7")
-                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.7")
-                            .WithNugetDependency("Microsoft.JSInterop.WebAssembly", "9.0.7"),
-                        ( >= 8, >= 0) => new PackageVersion("8.0.18")
-                            .WithNugetDependency("Microsoft.AspNetCore.Components.Web", "8.0.18")
+                        ( >= 9, >= 0) => new PackageVersion("9.0.8")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Web", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Configuration.Binder", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Configuration.Json", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.8")
+                            .WithNugetDependency("Microsoft.JSInterop.WebAssembly", "9.0.8"),
+                        ( >= 8, >= 0) => new PackageVersion("8.0.19")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Web", "8.0.19")
                             .WithNugetDependency("Microsoft.Extensions.Configuration.Binder", "8.0.2")
                             .WithNugetDependency("Microsoft.Extensions.Configuration.Json", "8.0.1")
                             .WithNugetDependency("Microsoft.Extensions.Logging", "8.0.1")
-                            .WithNugetDependency("Microsoft.JSInterop.WebAssembly", "8.0.18"),
+                            .WithNugetDependency("Microsoft.JSInterop.WebAssembly", "8.0.19"),
                         ( >= 7, >= 0) => new PackageVersion("7.0.20")
                             .WithNugetDependency("Microsoft.AspNetCore.Components.Web", "7.0.20")
                             .WithNugetDependency("Microsoft.Extensions.Configuration.Binder", "7.0.4")
@@ -50,12 +50,12 @@ namespace Intent.Modules.Blazor
             NugetRegistry.Register(MicrosoftAspNetCoreComponentsWebAssemblyAuthenticationPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 9, >= 0) => new PackageVersion("9.0.7")
-                            .WithNugetDependency("Microsoft.AspNetCore.Components.Authorization", "9.0.7")
-                            .WithNugetDependency("Microsoft.AspNetCore.Components.Web", "9.0.7"),
-                        ( >= 8, >= 0) => new PackageVersion("8.0.18")
-                            .WithNugetDependency("Microsoft.AspNetCore.Components.Authorization", "8.0.18")
-                            .WithNugetDependency("Microsoft.AspNetCore.Components.Web", "8.0.18"),
+                        ( >= 9, >= 0) => new PackageVersion("9.0.8")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Authorization", "9.0.8")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Web", "9.0.8"),
+                        ( >= 8, >= 0) => new PackageVersion("8.0.19")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Authorization", "8.0.19")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Web", "8.0.19"),
                         ( >= 7, >= 0) => new PackageVersion("7.0.20")
                             .WithNugetDependency("Microsoft.AspNetCore.Components.Authorization", "7.0.20")
                             .WithNugetDependency("Microsoft.AspNetCore.Components.Web", "7.0.20"),
@@ -71,8 +71,8 @@ namespace Intent.Modules.Blazor
             NugetRegistry.Register(MicrosoftAspNetCoreComponentsWebAssemblyServerPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 9, >= 0) => new PackageVersion("9.0.7"),
-                        ( >= 8, >= 0) => new PackageVersion("8.0.18"),
+                        ( >= 9, >= 0) => new PackageVersion("9.0.8"),
+                        ( >= 8, >= 0) => new PackageVersion("8.0.19"),
                         ( >= 7, >= 0) => new PackageVersion("7.0.20"),
                         ( >= 6, >= 0) => new PackageVersion("6.0.36"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{MicrosoftAspNetCoreComponentsWebAssemblyServerPackageName}'"),