+++ Modules/Intent.Modules.EntityFrameworkCore/NugetPackages.cs
@@ -352,9 +352,9 @@ namespace Intent.Modules.EntityFrameworkCore
             NugetRegistry.Register(PomeloEntityFrameworkCoreMySqlPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 8, >= 0) => new PackageVersion("8.0.3")
-                            .WithNugetDependency("Microsoft.EntityFrameworkCore.Relational", "8.0.13")
-                            .WithNugetDependency("MySqlConnector", "2.3.5"),
+                        ( >= 8, >= 0) => new PackageVersion("9.0.0")
+                            .WithNugetDependency("Microsoft.EntityFrameworkCore.Relational", "9.0.0")
+                            .WithNugetDependency("MySqlConnector", "2.4.0"),
                         ( >= 7, >= 0) => new PackageVersion("7.0.0"),
                         ( >= 6, >= 0) => new PackageVersion("7.0.0"),
                         ( >= 2, >= 1) => new PackageVersion("5.0.4")
@@ -371,12 +371,12 @@ namespace Intent.Modules.EntityFrameworkCore
             NugetRegistry.Register(PomeloEntityFrameworkCoreMySqlNetTopologySuitePackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 8, >= 0) => new PackageVersion("8.0.3")
-                            .WithNugetDependency("Microsoft.EntityFrameworkCore.Relational", "8.0.13")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "8.0.1")
-                            .WithNugetDependency("MySqlConnector", "2.3.5")
+                        ( >= 8, >= 0) => new PackageVersion("9.0.0")
+                            .WithNugetDependency("Microsoft.EntityFrameworkCore.Relational", "9.0.0")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "9.0.0")
+                            .WithNugetDependency("MySqlConnector", "2.4.0")
                             .WithNugetDependency("NetTopologySuite", "2.5.0")
-                            .WithNugetDependency("Pomelo.EntityFrameworkCore.MySql", "8.0.3"),
+                            .WithNugetDependency("Pomelo.EntityFrameworkCore.MySql", "9.0.0"),
                         ( >= 7, >= 0) => new PackageVersion("7.0.0"),
                         ( >= 6, >= 0) => new PackageVersion("7.0.0"),
                         ( >= 2, >= 1) => new PackageVersion("5.0.4")