+++ Modules/Intent.Modules.MongoDb.MongoFramework/NugetPackages.cs
@@ -19,15 +19,15 @@ namespace Intent.Modules.MongoDb.MongoFramework
             NugetRegistry.Register(FinbuckleMultiTenantPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 9, >= 0) => new PackageVersion("9.2.2")
-                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "9.0.7")
-                            .WithNugetDependency("Microsoft.Extensions.Configuration", "9.0.7")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "9.0.7")
-                            .WithNugetDependency("Microsoft.Extensions.Http", "9.0.7")
-                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.7")
-                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.7")
-                            .WithNugetDependency("Microsoft.Extensions.Options.ConfigurationExtensions", "9.0.7"),
-                        ( >= 8, >= 0) => new PackageVersion("9.2.2")
+                        ( >= 9, >= 0) => new PackageVersion("9.3.0")
+                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Configuration", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Http", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Options.ConfigurationExtensions", "9.0.8"),
+                        ( >= 8, >= 0) => new PackageVersion("9.3.0")
                             .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "8.0.0")
                             .WithNugetDependency("Microsoft.Extensions.Configuration", "8.0.0")
                             .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "8.0.1")
@@ -35,7 +35,7 @@ namespace Intent.Modules.MongoDb.MongoFramework
                             .WithNugetDependency("Microsoft.Extensions.Logging", "8.0.1")
                             .WithNugetDependency("Microsoft.Extensions.Options", "8.0.2")
                             .WithNugetDependency("Microsoft.Extensions.Options.ConfigurationExtensions", "8.0.0"),
-                        ( >= 7, >= 0) => new PackageVersion("8.1.5")
+                        ( >= 7, >= 0) => new PackageVersion("8.1.6")
                             .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "7.0.0")
                             .WithNugetDependency("Microsoft.Extensions.Configuration", "7.0.0")
                             .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "7.0.0")