+++ Modules/Intent.Modules.Blazor/release-notes.md
@@ -1,5 +1,6 @@
 ### Version 1.0.2
 
+- Improvement: Updated NuGet package versions.
 - Fixed: Blazor server local service resolved not handling remote services correctly.
 - Fixed: Corrected comparison of application Ids when determining the locality of a service invocation.