+++ Modules/Intent.Modules.Integration.HttpClients/release-notes.md
@@ -1,5 +1,6 @@
 ### Version 6.0.6
 
+- Improvement: Updated NuGet package versions.
 - Improvement: Removed dependency on `Intent.Modelers.ServiceProxies` module.
 
 ### Version 6.0.5