+++ Modules/Intent.Modules.MongoDb.MongoFramework/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.0.27
+
+- Improvement: Updated NuGet package versions.
+
 ### Version 1.0.26
 
 - Improvement: Updated NuGet package versions.